VFS: Roll out mount flag differentiation (MS_* -> SB_*) generally
The previous patch differentiated the MS_* flags passed to mount(2) from
the internal flags set in the super_block's s_flags.  s_flags are now
called SB_*, with the names and the values for the moment mirroring the
MS_* flags that they're equivalent to.

In this patch, roll things the conversion out generally.

Suggested-by: Al Viro <viro@ZenIV.linux.org.uk>
Signed-off-by: David Howells <dhowells@redhat.com>