mm, memory_hotplug: display allowed zones in the preferred ordering
Prior to "mm, memory_hotplug: do not associate hotadded memory to zones
until online" we used to allow to change the valid zone types of a memory
block if it is adjacent to a different zone type.  This fact was reflected
in memoryNN/valid_zones by the ordering of printed zones.  The first one
was default (echo online > memoryNN/state) and the other one could be
onlined explicitly by online_{movable,kernel}.  This behavior was removed
by the said patch and as such the ordering was not all that important.  In
most cases a kernel zone would be default anyway.  The only exception is
movable_node handled by "mm, memory_hotplug: support movable_node for
hotpluggable nodes".

Let's reintroduce this behavior again because later patch will remove the
zone overlap restriction and so user will be allowed to online kernel
resp.  movable block regardless of its placement.  Original behavior will
then become significant again because it would be non-trivial for users to
see what is the default zone to online into.

Implementation is really simple.  Pull out zone selection out of
move_pfn_range into zone_for_pfn_range helper and use it in
show_valid_zones to display the zone for default onlining and then both
kernel and movable if they are allowed.  Default online zone is not
duplicated.

Link: http://lkml.kernel.org/r/20170714121233.16861-2-mhocko@kernel.org
Signed-off-by: Michal Hocko <mhocko@suse.com>
Acked-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Acked-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Mel Gorman <mgorman@suse.de>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Reza Arbab <arbab@linux.vnet.ibm.com>
Cc: Yasuaki Ishimatsu <yasu.isimatu@gmail.com>
Cc: Xishi Qiu <qiuxishi@huawei.com>
Cc: Kani Toshimitsu <toshi.kani@hpe.com>
Cc: <slaoub@gmail.com>
Cc: Daniel Kiper <daniel.kiper@oracle.com>
Cc: Igor Mammedov <imammedo@redhat.com>
Cc: Vitaly Kuznetsov <vkuznets@redhat.com>
Cc: Wei Yang <richard.weiyang@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>