fscache: remove unused ->now_uncached callback
Patch series "Ranged pagevec lookup", v2.

In this series I make pagevec_lookup() update the index (to be consistent
with pagevec_lookup_tag() and also as a preparation for ranged lookups),
provide ranged variant of pagevec_lookup() and use it in places where it
makes sense.  This not only removes some common code but is also a
measurable performance win for some use cases (see patch 4/10) where radix
tree is sparse and searching & grabing of a page after the end of the
range has measurable overhead.

This patch (of 10):

The callback doesn't ever get called.  Remove it.

Link: http://lkml.kernel.org/r/20170726114704.7626-2-jack@suse.cz
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>