mm, oom: allow oom reaper to race with exit_mmap
David has noticed that the oom killer might kill additional tasks while
the exiting oom victim hasn't terminated yet because the oom_reaper marks
the curent victim MMF_OOM_SKIP too early when mm->mm_users dropped down to
0.  The race is as follows

oom_reap_task				do_exit
					  exit_mm
  __oom_reap_task_mm
					    mmput
					      __mmput
    mmget_not_zero # fails
    						exit_mmap # frees memory
  set_bit(MMF_OOM_SKIP)

The victim is still visible to the OOM killer until it is unhashed.

Currently we try to reduce a risk of this race by taking oom_lock and wait
for out_of_memory sleep while holding the lock to give the victim some
time to exit.  This is quite suboptimal approach because there is no
guarantee the victim (especially a large one) will manage to unmap its
address space and free enough memory to the particular oom domain which
needs a memory (e.g.  a specific NUMA node).

Fix this problem by allowing __oom_reap_task_mm and __mmput path to race.
__oom_reap_task_mm is basically MADV_DONTNEED and that is allowed to run
in parallel with other unmappers (hence the mmap_sem for read).

The only tricky part is to exclude page tables tear down and all
operations which modify the address space in the __mmput path.  exit_mmap
doesn't expect any other users so it doesn't use any locking.  Nothing
really forbids us to use mmap_sem for write, though.  In fact we are
already relying on this lock earlier in the __mmput path to synchronize
with ksm and khugepaged.

Take the exclusive mmap_sem when calling free_pgtables and destroying vmas
to sync with __oom_reap_task_mm which take the lock for read.  All other
operations can safely race with the parallel unmap.

The previous version of the patch has been posted here [1].  The original
patch has taken mmap_sem in exit_mmap unconditionally but Kirill was
worried this could have a performance impact (we should exercise the fast
path most of the time because nobody should be holding lock at that
stage).  An artificial testcase [2] has shown ~3% difference but numbers
are quite noisy [3] so it is effect is not all that clear.  Anyway I have
made the lock conditional for oom victims.

Andrea has proposed and alternative solution [4] which should be
equivalent functionally similar to {ksm,khugepaged}_exit.  I have to
confess I really don't like that approach but I can live with it if that
is a preferred way (to be honest I would like to drop the empty
down_write();up_write() from the other two callers as well).  In fact I
have asked Andrea to post his patch [5] but that hasn't happened.  I do
not think we should wait much longer and finally merge some fix.

[1] http://lkml.kernel.org/r/20170724072332.31903-1-mhocko@kernel.org
[2] http://lkml.kernel.org/r/20170725142626.GJ26723@dhcp22.suse.cz
[3] http://lkml.kernel.org/r/20170725160359.GO26723@dhcp22.suse.cz
[4] http://lkml.kernel.org/r/20170726162912.GA29716@redhat.com
[5] http://lkml.kernel.org/r/20170728062345.GA2274@dhcp22.suse.cz

Link: http://lkml.kernel.org/r/20170810081632.31265-1-mhocko@kernel.org
Fixes: 26db62f179d1 ("oom: keep mm of the killed task available")
Signed-off-by: Michal Hocko <mhocko@suse.com>
Reported-by: David Rientjes <rientjes@google.com>
Cc: Tetsuo Handa <penguin-kernel@I-love.SAKURA.ne.jp>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: Andrea Argangeli <andrea@kernel.org>
Cc: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill@shutemov.name>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>