MIPS: Introduce CPU_ISA_GE_* Kconfig entries
Various entries in Kconfig depend upon the ISA that the kernel build is
targeting being greater than or equal to some revision. The way in which
we've handled that so far has been to OR the Kconfig entries for all
applicable ISAs or CPUs. This leads to us having lists in Kconfig which:

 - Require expansion when new architecture revisions are introduced.

 - Don't always clearly show what we actually depend upon.

 - Make it difficult for code, or board Kconfig fragments in a later
   patch, to depend upon a range of ISA revisions in a maintainable way.

This patch introduces new Kconfig entries which indicate that the ISA
being targeted is greater than or equal to a particular revision, which
allows us to express requirements on potentially open-ended ranges of
ISA revisions rather than only upon specific revisions.

For example, we currently express the dependency of hardware watchpoint
support on a MIPSr1 or higher ISA like so:

  default y if CPU_MIPSR1 || CPU_MIPSR2 || CPU_MIPSR6

With the new Kconfig entries introduced by this patch this can be
simplified to:

  default y if CPU_ISA_GE_R1

Which makes it clearer what the actual dependency is & won't require
extending if we add further architecture revisions in the future (so
long as they support hardware watchpoints).

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Cc: James Hogan <james.hogan@imgtec.com>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/17060/
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>