x86/vector: Exclude IRQ0 from reservation mode
The timer check in the IOAPIC initialization code relies on the assumption
that the timer interrupt cannot move away from vector 48. This is required
to switch the timer interrupt delivery back to PIC mode when the delivery
through the IOAPIC fails.

Allowing IRQ0 to be part of the reservation mode violates that assumption
because the IOAPIC test must deactivate and reactivate the IOAPIC for pin
reconfiguration. If the interrupt is in reservation mode then it drops
vector 48 at deactivation and allocates the next free vector at activation,
which is unlikely to be vector 48. If the IOAPIC delivery fails, then the
timer IRQ is switched back to PIC mode which only works when IRQ0 is
delivered on vector 48. If the vector has moved, the timer does not work
and the kernel panics.

Remove the can_reserve property from the timer interrupt.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>