MAINTAINERS: Fix path and add bindings to timers
As spotted by Andreas Färber, the clocksource directory path does not follow
the rule where a maintained directory must end with a '\' character.

Also, the timers devicetree bindings documentation is not mentionned in the
entry, so every submission touching the devicetree documentation misses to
Cc the maintainers of the timers.

Reported-by: Andreas Färber <afaerber@suse.de>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>