clk: samsung: Add a separate driver for Exynos4412 ISP clocks
Some registers for the Exynos 4412 ISP (Camera subsystem) clocks are
partially located in the SOC area, which belongs to ISP power domain.
Because those registers are also located in a different memory region
than the main clock controller, support for them can be provided by
a separate clock controller. This in turn allows to almost seamlessly
make it aware of the power domain using recently introduced runtime
PM support for clocks.

Signed-off-by: Marek Szyprowski <m.szyprowski@samsung.com>
Acked-by: Krzysztof Kozlowski <krzk@kernel.org>
Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>