PCI: Create iov virtfn/physfn links before attaching driver
When creating virtual functions, create the "virtfn%u" and "physfn" links
in sysfs before attaching the driver.  Without this, there is a race when
the driver attaches to the new virtual network interface and sends out an
"add" udev event, and the network interface naming software (biosdevname or
systemd, for example) tries to look at these links.

Signed-off-by: Stuart Hayes <stuart.w.hayes@gmail.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>