usercopy: Allow for temporary fallback for non-whitelisted usercopy
While it is believed that all needed usercopy whitelist entries have been
identified and marked, it is possible some have been missed. Given how
long it took for the missing KVM whitelist to be noticed, it seems that
the usercopy whitelisting feature should continue to bake for a release
or more. As such, the new CONFIG_HARDENED_USERCOPY_FALLBACK is introduced
to allow usercopy operations for slabs that lack a whitelist, producing a
WARN instead of a full rejection. Once enough time passes with no one
tripping over a WARN, this option and its fallback code will be removed.

Signed-off-by: Kees Cook <keescook@chromium.org>