Unify migrate_pages and move_pages access checks
197e7e521384a23b ("Sanitize 'move_pages()' permission checks") fixed a
security issue I reported in the move_pages syscall, and made it so that
you can't act on set-uid processes unless you have the CAP_SYS_PTRACE
capability.

Unify the access check logic of migrate_pages to match the new behavior of
move_pages.  We discussed this a bit in the security@ list and thought
it'd be good for consistency even though there's no evident security
impact.  The NUMA node access checks are left intact and require
CAP_SYS_NICE as before.

Link: http://lkml.kernel.org/r/alpine.DEB.2.11.1710011830320.6333@lakka.kapsi.fi
Signed-off-by: Otto Ebeling <otto.ebeling@iki.fi>
Acked-by: Michal Hocko <mhocko@suse.com>
Cc: Eric W. Biederman <ebiederm@xmission.com>
Cc: Willy Tarreau <w@1wt.eu>
Cc: Kees Cook <keescook@chromium.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>