lib/rbtree-test: lower default params
Fengguang reported soft lockups while running the rbtree and interval tree
test modules.  The logic for these tests all occur in init phase, and we
currently are pounding with the default values for number of nodes and
number of iterations of each test.  Reduce the latter by two orders of
magnitude.  This does not influence the value of the tests in that one
thousand times by default is enough to get the picture.

Link: http://lkml.kernel.org/r/20171109161715.xai2dtwqw2frhkcm@linux-n805
Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Reported-by: Fengguang Wu <fengguang.wu@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>